+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 TBD: Version 1.1 (???)
 -------------------------------------------------------------------------------------------
+- #8: Fix Method metrics with wrong line number for annotated method.
 - #9: Add Gradle wrapper & remove IDE specific files.  (Dominik Broj)
 - #10: Refactorings: Remove deprecated org.codehaus.groovy.ast.expr.RegexExpression; remove duplicated defs of groovy's transform.powerassert.Value; add generics to MetricSet.getMetrics(). (Dominik Broj)